Fix Snyk client failing to encode PURLs properly
The Snyk API requires *all* special characters to be URL encoded. However, the MicroProfile `@RestClient` would not encode `@` and `:`. It's not possible to customize the URL encoding, so for the time being we're stuck with writing out own client logic.

Also re-added the handling of disabled analyzers. Analyzers that are not enabled will be part of the topology, and thus will not show up in the topology diagram.

Signed-off-by: nscuro <nscuro@protonmail.com>