Rough implementation of stateful retries for OSS Index
Implements the OSS Index analyzer as stream processor in order to have access to state stores. Because processors can only forward or drop records, they can't publish records to arbitrary channels, just the one they're connected to in the topology.

Incoming records are assembled into batches of up to 128. Batches are analyzed when their size exceeds 128, or 5 seconds since the last batch analysis have passed.

Records that failed with a "retryable reason" are written to a separate state store, with additional retry metadata. The retry store is checked every second for records that are due for retry. Records due for retry are re-added to the batch. The retry mechanism uses exponential backoff with jitter.

A circuit breaker is added so that retries do not overwhelm the OSS Index API. Per the current default, if 50% out of 100 requests to OSS Index fail, the circuit is opened for 15 seconds.

The `VulnerabilityResult` class has been extended with a `failureReason` field, that can be used to communicate *if* and *why* the analysis of a given record failed. Just a quick & dirty solution.

Cache is still missing from this implementation.

Signed-off-by: nscuro <nscuro@protonmail.com>