Implementation of stateful retries for vulnerability analyzers (#174)
* Rough implementation of stateful retries for OSS Index

Implements the OSS Index analyzer as stream processor in order to have access to state stores. Because processors can only forward or drop records, they can't publish records to arbitrary channels, just the one they're connected to in the topology.

Incoming records are assembled into batches of up to 128. Batches are analyzed when their size exceeds 128, or 5 seconds since the last batch analysis have passed.

Records that failed with a "retryable reason" are written to a separate state store, with additional retry metadata. The retry store is checked every second for records that are due for retry. Records due for retry are re-added to the batch. The retry mechanism uses exponential backoff with jitter.

A circuit breaker is added so that retries do not overwhelm the OSS Index API. Per the current default, if 50% out of 100 requests to OSS Index fail, the circuit is opened for 15 seconds.

The `VulnerabilityResult` class has been extended with a `failureReason` field, that can be used to communicate *if* and *why* the analysis of a given record failed. Just a quick & dirty solution.

Cache is still missing from this implementation.

Signed-off-by: nscuro <nscuro@protonmail.com>

* Move retry logic into abstract processor class

Signed-off-by: nscuro <nscuro@protonmail.com>

* Keep track of entry counts in retry state stores

Signed-off-by: nscuro <nscuro@protonmail.com>

* Remove unused processors

Signed-off-by: nscuro <nscuro@protonmail.com>

* Track number of entries in batch state stores for OSS Index

Signed-off-by: nscuro <nscuro@protonmail.com>

* Add initial test for SnykProcessor

Signed-off-by: nscuro <nscuro@protonmail.com>

* Completely migrate analyzers to processors

Also move analyzer configs into centralized classes, add configuration classes for dependencies of processors, add support for multiple Snyk tokens, make usage of batch state store for OSS Index more efficient, configure HTTP client timeouts for both OSS Index and Snyk.

Signed-off-by: nscuro <nscuro@protonmail.com>

* Make `RetryableRecord` extend `Record`

Makes treating of retryable records more seamless, and also retains metadata like headers.

Signed-off-by: nscuro <nscuro@protonmail.com>

* Unify retry config; Re-enable SnykClientTest

Signed-off-by: nscuro <nscuro@protonmail.com>

* Add OssIndexProcessor tests

Signed-off-by: nscuro <nscuro@protonmail.com>

* Migrate internal analyzer to processor

Signed-off-by: nscuro <nscuro@protonmail.com>

* Fix missing `@Transactional` context for internal scanner processor

Can't use `@Transactional` here as processors are not (and can not be) CDI-managed beans.

Signed-off-by: nscuro <nscuro@protonmail.com>

* Update processors for recent cache and completion changes

Signed-off-by: nscuro <nscuro@protonmail.com>

* Disable k8s dev services

It starts a cluster which we do not use

Signed-off-by: nscuro <nscuro@protonmail.com>

* Add tests for AuthorizationClientHeaderFactory

Signed-off-by: nscuro <nscuro@protonmail.com>

* Fix Vulnerability being marked as deprecated

Signed-off-by: nscuro <nscuro@protonmail.com>

* Report analyzer identity `NONE` for records with unknown identifiers

Signed-off-by: nscuro <nscuro@protonmail.com>

* Make sonar happy

Signed-off-by: nscuro <nscuro@protonmail.com>

* Fix test; Reduce retry log level to `debug`

Signed-off-by: nscuro <nscuro@protonmail.com>

* Port circuit breaker tests

Also populate circuit breaker configs with default values as per https://resilience4j.readme.io/docs/circuitbreaker#create-and-configure-a-circuitbreaker and remove the, now unused, CircuitBreakerCustom class.

Signed-off-by: nscuro <nscuro@protonmail.com>

* Update Snyk property names in docker-compose.yml

Signed-off-by: nscuro <nscuro@protonmail.com>

* Fix "No serializer found for class" exceptions when running as native image

Classes that are (de-)serialized with Jackson have to be explicitly registered for reflection, see https://quarkus.io/guides/writing-native-applications-tips#registering-for-reflection

Also fix `ClassNotFoundException: org.apache.commons.logging.impl.LogFactoryImpl`, see https://github.com/quarkusio/quarkus/issues/10128

Signed-off-by: nscuro <nscuro@protonmail.com>

* Fix broken merge artifacts

Signed-off-by: nscuro <nscuro@protonmail.com>

Signed-off-by: nscuro <nscuro@protonmail.com>