Fix breaking change in `NEW_VULNERABILITY` notification JSON format
Companion change to https://github.com/DependencyTrack/hyades-apiserver/pull/290

Note: Because the original field numbers and types remain untouched, this is not a breaking change schema-wise. The name of the fields has changed, but this is only relevant to the JSON output, which is exactly what we want to fix here.
Signed-off-by: nscuro <nscuro@protonmail.com>