Fix potential `TopologyException` when running multiple vulnerability-analyzer instances
Kafka Streams works based on topologies. It relies on the fact that the *exact* same topology is present on all instances of a KS app. This means that the order of topology nodes must be identical.

For the vulnerability-analyzer, users can enable or disable individual scanners. Only enabled scanners will be added to the topology on startup.

Scanners are CDI beans, and we inject all of them via [`Instance`](https://jakarta.ee/specifications/cdi/3.0/apidocs/jakarta/enterprise/inject/instance). `Instance` can contain zero or more beans.

Scanners are added to the topology by iterating over the `Instance` bean. The problem here is that the order of the beans in `Instance` is not guaranteed. It could happen that instance A of the service added scanners in the order 123, whereas instance B added them in order 213.

A symptom of this problem is the following exception, followed by a shutdown of Kafka Streams:

```
org.apache.kafka.streams.errors.TopologyException: Invalid topology: Topic hyades-vulnerability-analyzer-scan-task-ossindex-repartition is unknown to the topology. This may happen if different KafkaStreams instances of the same application execute different Topologies. Note that Topologies are only identical if all operators are added in the same order.
```

Sorting the beans before adding them to the topology resolved the issue.

Signed-off-by: nscuro <nscuro@protonmail.com>
Signed-off-by: mehab <meha.bhargava@citi.com>