Add script to generate config docs
The script parses a given `application.properties` file and generates a Markdown document for the properties within it. A custom Pebble template can be provided.

The following property annotations are supported:

* `@category`: Allows for categorization / grouping of related properties
* `@type`: Defines the type of the property
* `@default`: To be used for cases where the default value is implicit, for example when it is inherited from the framework or other properties
* `@example`: To give an idea of what a valid value may look like, when it's not possible to provide a sensible default value
* `@hidden`: Marks a property as to-be-excluded from the generated docs
* `@required`: Marks a property as required

We can easily add more as we see fit.

JBang is required to run the script: https://www.jbang.dev/documentation/guide/latest/installation.html

Closes #1154

Signed-off-by: nscuro <nscuro@protonmail.com>