Pull config via `@ConfigProperty` in mirror-service
Also removes manual retry logic, as the `open-vulnerability-clients` library has added its own retry since we originally adopted it.

Further removed the `num-threads` configuration option for the NVD datasource. Downloading of the data has never been a bottleneck, it's always the processing of already downloaded data that's the most time-consuming.

Signed-off-by: nscuro <nscuro@protonmail.com>