Generate documentation for Protobuf definitions
Adds basic documentation for `policy.proto` and `notification.proto` and makes them available under *Reference* -> *Schemas* in the docs site.

This is useful to link to the definitions from other areas in the docs.

A GitHub Actions workflow to update the docs automatically when the `.proto` files change is included.

Signed-off-by: nscuro <nscuro@protonmail.com>