Add initializer container to Docker Compose
The initializer executes schema migrations and populates the database with default objects. We added it to the Helm chart already (see https://github.com/DependencyTrack/helm-charts/issues/136), but it was still missing from `docker-compose.yml`.

The benefit of this addition is that we avoid any schema-related errors in hyades services, and the API server executing migrations and default object population on every startup.

Signed-off-by: nscuro <nscuro@protonmail.com>