Add upgrade notes
While we do label breaking changes in pull requests and issues, we can't expect users to sift through all of them to determine whether any action is needed upon upgrading to a newer version.

This adds a new Upgrading page to the documentation, which will list notable and breaking changes for each version.

Signed-off-by: nscuro <nscuro@protonmail.com>